+++ addons/taliban/subconfig_cup/CfgVehicles.hpp
@@ -1,173 +1,11 @@
 class CfgVehicles {
     #include "\x\tacu\addons\assets\script_classes_vehicles_cup.hpp"
 
-    // Militia
-    class TACU_Taliban_Militia_V_I_AGS30: TACU_Assets_AGS_base {
-        MACRO_FACTION_VEHICLE_DEFAULTS;
-        faction = "TACU_Taliban_Militia_I";
-        side = 2;
-        crew = "TACU_Taliban_Militia_U_I_Driver";
-        typicalCargo[] = {"TACU_Taliban_Militia_U_I_Driver"};
-    };
-    class TACU_Taliban_Militia_V_O_AGS30: TACU_Taliban_Militia_V_I_AGS30 {
-        faction = "TACU_Taliban_Militia_O";
-        side = 0;
-        crew = "TACU_Taliban_Militia_U_O_Driver";
-        typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
-    };
+    #include "CfgVehicles_Backpacks.hpp"
 
-    class TACU_Taliban_Militia_V_I_D30: TACU_Assets_D30_base {
-        MACRO_FACTION_VEHICLE_DEFAULTS;
-        faction = "TACU_Taliban_Militia_I";
-        side = 2;
-        crew = "TACU_Taliban_Militia_U_I_Driver";
-        typicalCargo[] = {"TACU_Taliban_Militia_U_I_Driver"};
-    };
-    class TACU_Taliban_Militia_V_O_D30: TACU_Taliban_Militia_V_I_D30 {
-        faction = "TACU_Taliban_Militia_O";
-        side = 0;
-        crew = "TACU_Taliban_Militia_U_O_Driver";
-        typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
-    };
+    #include "CfgVehicles_I.hpp"
+    #include "CfgVehicles_O.hpp"
 
-    class TACU_Taliban_Militia_V_I_DSHKM: TACU_Assets_DSHKM_base {
-        MACRO_FACTION_VEHICLE_DEFAULTS;
-        faction = "TACU_Taliban_Militia_I";
-        side = 2;
-        crew = "TACU_Taliban_Militia_U_I_Driver";
-        typicalCargo[] = {"TACU_Taliban_Militia_U_I_Driver"};
-    };
-    class TACU_Taliban_Militia_V_O_DSHKM: TACU_Taliban_Militia_V_I_DSHKM {
-        faction = "TACU_Taliban_Militia_O";
-        side = 0;
-        crew = "TACU_Taliban_Militia_U_O_Driver";
-        typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
-    };
-
-    class TACU_Taliban_Militia_V_I_DSHKM_Mini: TACU_Assets_DSHKM_Mini_base {
-        MACRO_FACTION_VEHICLE_DEFAULTS;
-        faction = "TACU_Taliban_Militia_I";
-        side = 2;
-        crew = "TACU_Taliban_Militia_U_I_Driver";
-        typicalCargo[] = {"TACU_Taliban_Militia_U_I_Driver"};
-    };
-    class TACU_Taliban_Militia_V_O_DSHKM_Mini: TACU_Taliban_Militia_V_I_DSHKM_Mini {
-        faction = "TACU_Taliban_Militia_O";
-        side = 0;
-        crew = "TACU_Taliban_Militia_U_O_Driver";
-        typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
-    };
-
-    class TACU_Taliban_Militia_V_I_KORD: TACU_Assets_KORD_base {
-        MACRO_FACTION_VEHICLE_DEFAULTS;
-        faction = "TACU_Taliban_Militia_I";
-        side = 2;
-        crew = "TACU_Taliban_Militia_U_I_Driver";
-        typicalCargo[] = {"TACU_Taliban_Militia_U_I_Driver"};
-    };
-    class TACU_Taliban_Militia_V_O_KORD: TACU_Taliban_Militia_V_I_KORD {
-        faction = "TACU_Taliban_Militia_O";
-        side = 0;
-        crew = "TACU_Taliban_Militia_U_O_Driver";
-        typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
-    };
-
-    class TACU_Taliban_Militia_V_I_KORD_Mini: TACU_Assets_KORD_Mini_base {
-        MACRO_FACTION_VEHICLE_DEFAULTS;
-        faction = "TACU_Taliban_Militia_I";
-        side = 2;
-        crew = "TACU_Taliban_Militia_U_I_Driver";
-        typicalCargo[] = {"TACU_Taliban_Militia_U_I_Driver"};
-    };
-    class TACU_Taliban_Militia_V_O_KORD_Mini: TACU_Taliban_Militia_V_I_KORD_Mini {
-        faction = "TACU_Taliban_Militia_O";
-        side = 0;
-        crew = "TACU_Taliban_Militia_U_O_Driver";
-        typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
-    };
-
-    // Veterans
-    class TACU_Taliban_Veterans_V_I_AGS30: TACU_Assets_AGS_base {
-        MACRO_FACTION_VEHICLE_DEFAULTS;
-        faction = "TACU_Taliban_Veterans_I";
-        side = 2;
-        crew = "TACU_Taliban_Veterans_U_I_Driver";
-        typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
-    };
-    class TACU_Taliban_Veterans_V_O_AGS30: TACU_Taliban_Veterans_V_I_AGS30 {
-        faction = "TACU_Taliban_Veterans_O";
-        side = 0;
-        crew = "TACU_Taliban_Veterans_U_O_Driver";
-        typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
-    };
-
-    class TACU_Taliban_Veterans_V_I_D30: TACU_Assets_D30_base {
-        MACRO_FACTION_VEHICLE_DEFAULTS;
-        faction = "TACU_Taliban_Veterans_I";
-        side = 2;
-        crew = "TACU_Taliban_Veterans_U_I_Driver";
-        typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
-    };
-    class TACU_Taliban_Veterans_V_O_D30: TACU_Taliban_Veterans_V_I_D30 {
-        faction = "TACU_Taliban_Veterans_O";
-        side = 0;
-        crew = "TACU_Taliban_Veterans_U_O_Driver";
-        typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
-    };
-
-    class TACU_Taliban_Veterans_V_I_DSHKM: TACU_Assets_DSHKM_base {
-        MACRO_FACTION_VEHICLE_DEFAULTS;
-        faction = "TACU_Taliban_Veterans_I";
-        side = 2;
-        crew = "TACU_Taliban_Veterans_U_I_Driver";
-        typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
-    };
-    class TACU_Taliban_Veterans_V_O_DSHKM: TACU_Taliban_Veterans_V_I_DSHKM {
-        faction = "TACU_Taliban_Veterans_O";
-        side = 0;
-        crew = "TACU_Taliban_Veterans_U_O_Driver";
-        typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
-    };
-
-    class TACU_Taliban_Veterans_V_I_DSHKM_Mini: TACU_Assets_DSHKM_Mini_base {
-        MACRO_FACTION_VEHICLE_DEFAULTS;
-        faction = "TACU_Taliban_Veterans_I";
-        side = 2;
-        crew = "TACU_Taliban_Veterans_U_I_Driver";
-        typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
-    };
-    class TACU_Taliban_Veterans_V_O_DSHKM_Mini: TACU_Taliban_Veterans_V_I_DSHKM_Mini {
-        faction = "TACU_Taliban_Veterans_O";
-        side = 0;
-        crew = "TACU_Taliban_Veterans_U_O_Driver";
-        typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
-    };
-
-    class TACU_Taliban_Veterans_V_I_KORD: TACU_Assets_KORD_base {
-        MACRO_FACTION_VEHICLE_DEFAULTS;
-        faction = "TACU_Taliban_Veterans_I";
-        side = 2;
-        crew = "TACU_Taliban_Veterans_U_I_Driver";
-        typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
-    };
-    class TACU_Taliban_Veterans_V_O_KORD: TACU_Taliban_Veterans_V_I_KORD {
-        faction = "TACU_Taliban_Veterans_O";
-        side = 0;
-        crew = "TACU_Taliban_Veterans_U_O_Driver";
-        typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
-    };
-
-    class TACU_Taliban_Veterans_V_I_KORD_Mini: TACU_Assets_KORD_Mini_base {
-        MACRO_FACTION_VEHICLE_DEFAULTS;
-        faction = "TACU_Taliban_Veterans_I";
-        side = 2;
-        crew = "TACU_Taliban_Veterans_U_I_Driver";
-        typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
-    };
-    class TACU_Taliban_Veterans_V_O_KORD_Mini: TACU_Taliban_Veterans_V_I_KORD_Mini {
-        faction = "TACU_Taliban_Veterans_O";
-        side = 0;
-        crew = "TACU_Taliban_Veterans_U_O_Driver";
-        typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
-    };
+    #include "CfgVehicles_Turrets_I.hpp"
+    #include "CfgVehicles_Turrets_O.hpp"
 };
