+++ addons/taliban/subconfig_cup/CfgVehicles_Backpacks.hpp
@@ -1,0 +1,7 @@
+class B_FieldPack_khk;
+class TACU_Taliban_Militia_B_AP: B_FieldPack_khk {
+    MACRO_BACKPACK_DEFAULTS;
+    class TransportMagazines {
+        MACRO_ADDMAGAZINE(CUP_OG7_M,3);
+    };
+};
