+++ addons/taliban/subconfig_cup/CfgVehicles_I.hpp
@@ -1,0 +1,31 @@
+class TACU_Taliban_Militia_U_I_AT;
+class TACU_Taliban_Militia_U_I_AP: TACU_Taliban_Militia_U_I_AT {
+    displayName = "Rifleman (AP)";
+    backpack = "TACU_Taliban_Militia_B_AP";
+    magazines[] = {
+        "CUP_OG7_M",
+        mag_12("TACU_Magazine_30Rnd_762_AK"),
+        mag_2("SmokeShell")
+    };
+    respawnMagazines[] = {
+        "CUP_OG7_M",
+        mag_12("TACU_Magazine_30Rnd_762_AK"),
+        mag_2("SmokeShell")
+    };
+};
+
+class TACU_Taliban_Veterans_U_I_AT;
+class TACU_Taliban_Veterans_U_I_AP: TACU_Taliban_Veterans_U_I_AT {
+    displayName = "Rifleman (AP)";
+    backpack = "TACU_Taliban_Militia_B_AP";
+    magazines[] = {
+        "CUP_OG7_M",
+        mag_12("TACU_Magazine_30Rnd_762_AK"),
+        mag_2("SmokeShell")
+    };
+    respawnMagazines[] = {
+        "CUP_OG7_M",
+        mag_12("TACU_Magazine_30Rnd_762_AK"),
+        mag_2("SmokeShell")
+    };
+};
