+++ addons/taliban/subconfig_cup/CfgVehicles_O.hpp
@@ -1,0 +1,8 @@
+class TACU_Taliban_Militia_U_O_AP: TACU_Taliban_Militia_U_I_AP {
+    side = 0;
+    faction = "TACU_Taliban_Militia_O";
+};
+class TACU_Taliban_Veterans_U_O_AP: TACU_Taliban_Veterans_U_I_AP {
+    side = 0;
+    faction = "TACU_Taliban_Veterans_O";
+};
