+++ addons/taliban/subconfig_cup/CfgVehicles_Turrets_I.hpp
@@ -1,0 +1,97 @@
+// Militia
+class TACU_Taliban_Militia_V_I_AGS30: TACU_Assets_AGS_base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Taliban_Militia_I";
+    side = 2;
+    crew = "TACU_Taliban_Militia_U_I_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_I_Driver"};
+};
+
+class TACU_Taliban_Militia_V_I_D30: TACU_Assets_D30_base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Taliban_Militia_I";
+    side = 2;
+    crew = "TACU_Taliban_Militia_U_I_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_I_Driver"};
+};
+
+class TACU_Taliban_Militia_V_I_DSHKM: TACU_Assets_DSHKM_base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Taliban_Militia_I";
+    side = 2;
+    crew = "TACU_Taliban_Militia_U_I_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_I_Driver"};
+};
+
+class TACU_Taliban_Militia_V_I_DSHKM_Mini: TACU_Assets_DSHKM_Mini_base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Taliban_Militia_I";
+    side = 2;
+    crew = "TACU_Taliban_Militia_U_I_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_I_Driver"};
+};
+
+class TACU_Taliban_Militia_V_I_KORD: TACU_Assets_KORD_base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Taliban_Militia_I";
+    side = 2;
+    crew = "TACU_Taliban_Militia_U_I_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_I_Driver"};
+};
+
+class TACU_Taliban_Militia_V_I_KORD_Mini: TACU_Assets_KORD_Mini_base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Taliban_Militia_I";
+    side = 2;
+    crew = "TACU_Taliban_Militia_U_I_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_I_Driver"};
+};
+
+// Veterans
+class TACU_Taliban_Veterans_V_I_AGS30: TACU_Assets_AGS_base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Taliban_Veterans_I";
+    side = 2;
+    crew = "TACU_Taliban_Veterans_U_I_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
+};
+
+class TACU_Taliban_Veterans_V_I_D30: TACU_Assets_D30_base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Taliban_Veterans_I";
+    side = 2;
+    crew = "TACU_Taliban_Veterans_U_I_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
+};
+
+class TACU_Taliban_Veterans_V_I_DSHKM: TACU_Assets_DSHKM_base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Taliban_Veterans_I";
+    side = 2;
+    crew = "TACU_Taliban_Veterans_U_I_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
+};
+
+class TACU_Taliban_Veterans_V_I_DSHKM_Mini: TACU_Assets_DSHKM_Mini_base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Taliban_Veterans_I";
+    side = 2;
+    crew = "TACU_Taliban_Veterans_U_I_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
+};
+
+class TACU_Taliban_Veterans_V_I_KORD: TACU_Assets_KORD_base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Taliban_Veterans_I";
+    side = 2;
+    crew = "TACU_Taliban_Veterans_U_I_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
+};
+
+class TACU_Taliban_Veterans_V_I_KORD_Mini: TACU_Assets_KORD_Mini_base {
+    MACRO_FACTION_VEHICLE_DEFAULTS;
+    faction = "TACU_Taliban_Veterans_I";
+    side = 2;
+    crew = "TACU_Taliban_Veterans_U_I_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_I_Driver"};
+};
