+++ addons/taliban/subconfig_cup/CfgVehicles_Turrets_O.hpp
@@ -1,0 +1,82 @@
+// Militia
+class TACU_Taliban_Militia_V_O_AGS30: TACU_Taliban_Militia_V_I_AGS30 {
+    faction = "TACU_Taliban_Militia_O";
+    side = 0;
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
+class TACU_Taliban_Militia_V_O_D30: TACU_Taliban_Militia_V_I_D30 {
+    faction = "TACU_Taliban_Militia_O";
+    side = 0;
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
+class TACU_Taliban_Militia_V_O_DSHKM: TACU_Taliban_Militia_V_I_DSHKM {
+    faction = "TACU_Taliban_Militia_O";
+    side = 0;
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
+
+class TACU_Taliban_Militia_V_O_DSHKM_Mini: TACU_Taliban_Militia_V_I_DSHKM_Mini {
+    faction = "TACU_Taliban_Militia_O";
+    side = 0;
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
+
+class TACU_Taliban_Militia_V_O_KORD: TACU_Taliban_Militia_V_I_KORD {
+    faction = "TACU_Taliban_Militia_O";
+    side = 0;
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
+class TACU_Taliban_Militia_V_O_KORD_Mini: TACU_Taliban_Militia_V_I_KORD_Mini {
+    faction = "TACU_Taliban_Militia_O";
+    side = 0;
+    crew = "TACU_Taliban_Militia_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Militia_U_O_Driver"};
+};
+
+// Veterans
+class TACU_Taliban_Veterans_V_O_AGS30: TACU_Taliban_Veterans_V_I_AGS30 {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+
+class TACU_Taliban_Veterans_V_O_D30: TACU_Taliban_Veterans_V_I_D30 {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+
+class TACU_Taliban_Veterans_V_O_DSHKM: TACU_Taliban_Veterans_V_I_DSHKM {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+
+class TACU_Taliban_Veterans_V_O_DSHKM_Mini: TACU_Taliban_Veterans_V_I_DSHKM_Mini {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+
+class TACU_Taliban_Veterans_V_O_KORD: TACU_Taliban_Veterans_V_I_KORD {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
+
+class TACU_Taliban_Veterans_V_O_KORD_Mini: TACU_Taliban_Veterans_V_I_KORD_Mini {
+    faction = "TACU_Taliban_Veterans_O";
+    side = 0;
+    crew = "TACU_Taliban_Veterans_U_O_Driver";
+    typicalCargo[] = {"TACU_Taliban_Veterans_U_O_Driver"};
+};
